+++ src/com/aixcoder/extension/jobs/AiXInsertUIJob.java
@@ -1,8 +1,13 @@
 package com.aixcoder.extension.jobs;
 
+import java.util.HashMap;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 
 import org.eclipse.jface.text.IRegion;
 import org.eclipse.jface.text.ITextViewer;
@@ -78,33 +83,44 @@ public class AiXInsertUIJob extends AiXUIJob {
 						break;
 					}
 				}
-				fSorter.list = predictResult.sortResults;
+				if (predictResult.sortResults != null) {
+					Map<ICompletionProposal, Double> scoreMap = new HashMap<ICompletionProposal, Double>();
+					for (SortResult sortResult : predictResult.sortResults) {
+						boolean matched = false;
+						for (ICompletionProposal p : fComputedProposal) {
+							String s = p.getDisplayString();
+							if (s.equals(sortResult.word) || s.startsWith(sortResult.word + " ")
+									|| s.startsWith(sortResult.word + "(")) {
+								matched = true;
+								scoreMap.put(p, sortResult.prob);
+								break;
+							}
+						}
+						if (sortResult.options != null && sortResult.options.forced && !matched) {
+							ICompletionProposal forcedProposal = proposalFactory.createForcedSortProposal(selection.x,
+									sortResult.word, predictResult.current, sortResult.prob);
+							fFilteredProposals.add(0, forcedProposal);
+							fComputedProposal.add(0, forcedProposal);
+							scoreMap.put(forcedProposal, sortResult.prob);
+						}
+					}
+					ArrayList<Entry<ICompletionProposal, Double>> sortedResults = new ArrayList<Entry<ICompletionProposal, Double>>(scoreMap.entrySet());
+					Collections.sort(sortedResults, new Comparator<Entry<ICompletionProposal, Double>>() {
+						@Override
+						public int compare(Entry<ICompletionProposal, Double> o1, Entry<ICompletionProposal, Double> o2) {
+							return (int) Math.signum(o2.getValue() - o1.getValue());
+						}
+					});
+					for (int i = 0; i < sortedResults.size(); i++) {
+						sortedResults.get(i).setValue((double) (sortedResults.size() - i));
+					}
+					fSorter.scoreMap = scoreMap;
+				}
 				if (fSorter.longProposal == null
 						&& (predictResult.sortResults == null || predictResult.sortResults.length == 0
 								|| !rendered.display.matches("[a-zA-Z0-9_$]+(\\()?"))) {
 					fFilteredProposals.add(0, proposal);
 					fComputedProposal.add(0, proposal);
-				}
-				if (predictResult.sortResults != null) {
-					for (SortResult sortResult : predictResult.sortResults) {
-						if (sortResult.options != null && sortResult.options.forced) {
-							boolean matched = false;
-							for (ICompletionProposal p : fComputedProposal) {
-								String s = p.getDisplayString();
-								if (s.equals(sortResult.word) || s.startsWith(sortResult.word + " ")
-										|| s.startsWith(sortResult.word + "(")) {
-									matched = true;
-									break;
-								}
-							}
-							if (!matched) {
-								ICompletionProposal forcedProposal = proposalFactory.createForcedSortProposal(
-										selection.x, sortResult.word, predictResult.current, sortResult.prob);
-								fFilteredProposals.add(0, forcedProposal);
-								fComputedProposal.add(0, forcedProposal);
-							}
-						}
-					}
 				}
 
 				new AiXReportJob(ReportType.SHOW).schedule();