+++ src/com/aixcoder/i18n/EN.java
@@ -18,6 +18,11 @@ public class EN extends Localization {
 		m.put(additionalParameters, "Additional &Parameters");
 		m.put(description,
 				"AiXcoder is an AI-powered code completion service. Visit https://aixcoder.com for more information.");
+		m.put(telemetryTitle, "aiXcoder user statistics collection");
+		m.put(telemetryQuestion, "Are you willing to send anonymous usage data to improve user experience? You can later change it in preferences page.");
+		m.put(endpointEmptyTitle, "aiXcoder endpoint is empty!");
+		m.put(endpointEmptyWarning, "The endpoint of aiXcoder is not set. Please set it manually in Window->Preferences->AiXCoder Preferences. Our public endpoint is https://api.aixcoder.com/");
+		m.put(longResultRank, "The preferred position of the long result");
 	}
 
 	public static String R(String input) {