+++ src/com/aixcoder/lib/Preference.java
@@ -19,8 +19,10 @@ public class Preference {
 	public static final String AUTO_IMPORT = "AUTO_IMPORT";
 	public static final String SORT_ONLY = "SORT_ONLY";
 	public static final String LANGUAGE = "LANGUAGE";
+	public static final String ASKED_LANGUAGE = "ASKED_LANGUAGE";
 	public static final String ALLOW_TELEMETRY = "ALLOW_TELEMETRY";
 	public static final String ASKED_TELEMETRY = "ASKED_TELEMETRY";
+	public static final String LONG_RESULT_RANK = "LONG_RESULT_RANK";
 
 	public static final String id = Activator.PLUGIN_ID + ".preferences.page";
 	public static ScopedPreferenceStore preferenceManager = new ScopedPreferenceStore(InstanceScope.INSTANCE, id);
@@ -77,5 +79,13 @@ public class Preference {
 
 	public static String getLanguage() {
 		return preferenceManager.getString(LANGUAGE);
+	}
+
+	public static boolean askedLanguage() {
+		return preferenceManager.getBoolean(ASKED_LANGUAGE);
+	}
+
+	public static int getLongResultRank() {
+		return preferenceManager.getInt(LONG_RESULT_RANK);
 	}
 }